Get the new theme without modifying the list
The new theme is now chosen without modifying the original
list of themes like before.

The new method gets the new theme by visiting the next item
in the list(or wrapping around to the beginning if the old theme
happens to be the last in the list).

A nice side effect is that the list of themes to swap can now
contain more than two themes.